+++ app/android/app/src/main/kotlin/com/kirana/kirana_gpt/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.kirana.kirana_gpt
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()